Refactor image upload logic from ChatInput to Chat component
