+++ packages/react/src/components/ChatInput.tsx
@@ -1,4 +1,4 @@
-import React, { useRef, useEffect, useState, useCallback } from 'react';
+import React, { useRef, useEffect, useCallback } from 'react';
 import { Send, Square, ImageIcon, X } from 'lucide-react';
 import { DistriPart } from '@distri/core';
 
@@ -18,6 +18,10 @@ export interface ChatInputProps {
   disabled?: boolean;
   isStreaming?: boolean;
   className?: string;
+  // Image upload props
+  attachedImages?: AttachedImage[];
+  onRemoveImage?: (id: string) => void;
+  onAddImages?: (files: FileList | File[]) => void;
 }
 
 export const ChatInput: React.FC<ChatInputProps> = ({
@@ -29,11 +33,12 @@ export const ChatInput: React.FC<ChatInputProps> = ({
   disabled = false,
   isStreaming = false,
   className = "",
+  attachedImages = [],
+  onRemoveImage,
+  onAddImages,
 }) => {
   const textareaRef = useRef<HTMLTextAreaElement>(null);
   const fileInputRef = useRef<HTMLInputElement>(null);
-  const [attachedImages, setAttachedImages] = useState<AttachedImage[]>([]);
-  const [isDragOver, setIsDragOver] = useState(false);
 
   // Auto-resize textarea
   useEffect(() => {
@@ -60,65 +65,14 @@ export const ChatInput: React.FC<ChatInputProps> = ({
     });
   }, []);
 
-  const addImages = useCallback(async (files: FileList | File[]) => {
-    const imageFiles = Array.from(files).filter(file => file.type.startsWith('image/'));
-
-    for (const file of imageFiles) {
-      const id = Date.now().toString() + Math.random().toString(36).substr(2, 9);
-      const preview = URL.createObjectURL(file);
-
-      const newImage: AttachedImage = {
-        id,
-        file,
-        preview,
-        name: file.name
-      };
-
-      setAttachedImages(prev => [...prev, newImage]);
-    }
-  }, []);
-
-  const removeImage = useCallback((id: string) => {
-    setAttachedImages(prev => {
-      const image = prev.find(img => img.id === id);
-      if (image) {
-        URL.revokeObjectURL(image.preview);
-      }
-      return prev.filter(img => img.id !== id);
-    });
-  }, []);
-
-  const handleDragOver = useCallback((e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragOver(true);
-  }, []);
-
-  const handleDragLeave = useCallback((e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragOver(false);
-  }, []);
-
-  const handleDrop = useCallback((e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragOver(false);
-
-    const files = e.dataTransfer.files;
-    if (files.length > 0) {
-      addImages(files);
-    }
-  }, [addImages]);
-
   const handleFileSelect = useCallback((e: React.ChangeEvent<HTMLInputElement>) => {
     const files = e.target.files;
-    if (files && files.length > 0) {
-      addImages(files);
+    if (files && files.length > 0 && onAddImages) {
+      onAddImages(files);
     }
     // Reset input value so same file can be selected again
     e.target.value = '';
-  }, [addImages]);
+  }, [onAddImages]);
 
   const handleKeyPress = (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
     if (e.key === 'Enter' && !e.shiftKey) {
@@ -163,12 +117,8 @@ export const ChatInput: React.FC<ChatInputProps> = ({
         onSend(value);
       }
 
-      // Clear input and images
+      // Clear input (images will be cleared by parent)
       onChange('');
-      setAttachedImages(prev => {
-        prev.forEach(img => URL.revokeObjectURL(img.preview));
-        return [];
-      });
 
     } catch (error) {
       console.error('Error processing images:', error);
@@ -184,13 +134,6 @@ export const ChatInput: React.FC<ChatInputProps> = ({
   const hasContent = value.trim().length > 0 || attachedImages.length > 0;
   const isDisabled = disabled || isStreaming;
 
-  // Clean up object URLs on unmount
-  useEffect(() => {
-    return () => {
-      attachedImages.forEach(img => URL.revokeObjectURL(img.preview));
-    };
-  }, []);
-
   return (
     <div className={`relative flex min-h-14 w-full items-end ${className}`}>
       <div className="relative flex w-full flex-auto flex-col">
@@ -205,7 +148,7 @@ export const ChatInput: React.FC<ChatInputProps> = ({
                   className="w-16 h-16 object-cover rounded-lg border border-border"
                 />
                 <button
-                  onClick={() => removeImage(image.id)}
+                  onClick={() => onRemoveImage?.(image.id)}
                   className="absolute -top-1 -right-1 w-5 h-5 bg-destructive hover:bg-destructive/90 text-destructive-foreground rounded-full flex items-center justify-center opacity-0 group-hover:opacity-100 transition-opacity"
                 >
                   <X className="w-3 h-3" />
@@ -218,13 +161,7 @@ export const ChatInput: React.FC<ChatInputProps> = ({
           </div>
         )}
 
-        <div
-          className={`relative mx-5 flex min-h-14 flex-auto rounded-lg border border-input bg-input items-start h-full ${isDragOver ? 'border-primary border-2 bg-primary/5' : ''
-            }`}
-          onDragOver={handleDragOver}
-          onDragLeave={handleDragLeave}
-          onDrop={handleDrop}
-        >
+        <div className="relative mx-5 flex min-h-14 flex-auto rounded-lg border border-input bg-input items-start h-full">
           <textarea
             ref={textareaRef}
             value={value}
@@ -235,13 +172,6 @@ export const ChatInput: React.FC<ChatInputProps> = ({
             rows={1}
             className="max-h-[25dvh] flex-1 resize-none border-none outline-none bg-transparent placeholder:text-muted-foreground focus:ring-0 overflow-auto text-sm p-4 pr-24 text-foreground min-h-[52px] max-h-[120px]"
           />
-
-          {/* Drag overlay */}
-          {isDragOver && (
-            <div className="absolute inset-0 flex items-center justify-center bg-primary/10 rounded-lg">
-              <div className="text-primary font-medium">Drop images here</div>
-            </div>
-          )}
 
           <div className="absolute right-2 bottom-0 flex items-center gap-1 h-full">
             {/* Image upload button */}
